requirements.txt
accelerate>=0.20.0
asyncio>=3.4.3
datasets>=2.12.0
diffusers>=0.23.0
faiss-cpu>=1.7.4
ftfy>=6.1.0
matplotlib>=3.7.0
numpy>=1.24.0
onnx>=1.14.0
onnxruntime>=1.15.0
opencv-python>=4.7.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
pillow>=9.0.0
prometheus-client>=0.17.0
protobuf>=4.23.0
psycopg2>=2.9.0
pydantic>=1.10.0
ray>=2.5.0
redis>=4.5.0
regex>=2023.5.5
safetensors>=0.3.1
scikit-image>=0.20.0
scipy>=1.10.0
sentencepiece>=0.1.97
tensorboard>=2.12.0
tokenizers>=0.13.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
transformers>=4.30.0
websockets>=11.0.0 